GUI work
This is a rather big leap in progress, following on from getting the database login to work, in this commit, the system now has a main interface which is linked with buttons to three other panels, all with the buttons created to provide a pop up for the action that it will perform and to provide basic navigation functionality to the application, such as going back and switching panels. There were also a few layout alterations trying to get things abit more in line and more appealing on the eye. The login still needs to be tidied up looks wise.