Display List in Stock Panel & Call Stock to List
This sprint led to a few issues, the UI of the booking panel has become somewhat contorted when the list had more than a single characteristic of a booking added. Both panels now have a functioning back button. Issues with GBC appearing with the way that i have managed to get the lists to read data, this may lead to a change in layout type for these panels.